QTPyFlashUtility-main/TelosAirSAMDBoardFlashGUI/requirements.txt
psutil
Pillow
pyduinocli
pycryptodome
pyserial